+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - core fix K.Util.sanitize.fileExt()
 - core K.Util.sanitize.name() lower param
 - core index 2dsphere geometry field
+- core new method K.Util.valid.point()
 - plugin geoinfo autoupdate geoinfo field when loc is changed
 - plugin upload rename param 'callback' in 'onUploaded' in template inputFile_upload
 - plugin upload new param 'params' in template inputFile_upload